+++ src/services/Renderer.php
@@ -22,31 +22,50 @@ use yii\base\Component;
 
 class Renderer extends Component
 {
+	const ELEMENT_TYPE_MAP = [
+		'craft\elements\Entry' => 'getEntryElements',
+		'craft\elements\GlobalSet' => 'getGlobalSetElements',
+		'craft\elements\Category' => 'getCategoryElements',
+		'craft\elements\Tag' => 'getTagElements',
+		'craft\elements\Asset' => 'getAssetElements',
+		'craft\elements\User' => 'getUserElements',
+		'craft\commerce\elements\Product' => 'getProductElements',
+		'craft\commerce\elements\Variant' => 'getVariantElements',
+	];
+
 	/**
 	 * @deprecated
 	 */
 	public function render(int $elementid, int $siteid)
 	{
-		Craft::$app->getDeprecator()->log('charliedev\elementmap\services\Renderer::render()', 'charliedev\elementmap\services\Renderer::render() is deprecated. Map results may be gathered individually with other functions such as getElementMap, getIncomingElements, and getOutgoingElements.');
+		Craft::$app->getDeprecator()->log(
+			'charliedev\elementmap\services\Renderer::render()',
+			'charliedev\elementmap\services\Renderer::render() is deprecated. Map results may be gathered individually with other functions such as getElementMap, getIncomingElements, and getOutgoingElements.'
+		);
 
 		// Gather up necessary structure data to render the element map with.
 		$results = $this->getElementMap($elementid, $siteid);
 
 		// Render the actual element map.
-		return Craft::$app->view->renderTemplate('element-map/_map', ['map' => $results]);
+		return Craft::$app->view->renderTemplate(
+			'element-map/_map',
+			[
+				'map' => $results
+			]
+		);
 	}
 
 	/**
 	 * Generates a data structure containing elements that reference the given
 	 * element and those that the given element references.
-	 * @param int $elementid The ID of the element to retrieve map
+	 * @param int $elementId The ID of the element to retrieve map
 	 * information about.
 	 * @param int $siteId The ID of the site context that information should
 	 * be gathered within.
 	 */
 	public function getElementMap($elementId, int $siteId)
 	{
-		if (!$elementId) {
+		if (!$elementId) { // No element, no element map.
 			return null;
 		}
 
@@ -58,61 +77,101 @@ class Renderer extends Component
 
 	/**
 	 * Retrieves a list of elements referencing the given element.
-	 * @param int $elementid The ID of the element to retrieve map
+	 * @param int $elementId The ID of the element to retrieve map
 	 * information about.
 	 * @param int $siteId The ID of the site context that information should
 	 * be gathered within.
 	 */
 	public function getIncomingElements($elementId, int $siteId)
 	{
-		if (!$elementId) {
+		if (!$elementId) { // No element, no related elements.
 			return null;
 		}
 
-		// Find incoming relationships to this element. Check for references to it, then trace those elements'
-		// owners until we get to meaningful things, such as Category -in-> Matrix Block -in-> Element.
-		$variants = $this->getVariantsByProduct($elementId); // Element -> variants.
-		$fromdata = $this->getRelationshipGroups(array_merge([$elementId], $variants), $siteId, true);
+		// Assemble a set of elements that should be used as the targets.
 
-		// Convert the retrieved element ids into data we use to display the map.
-		$this->processRelationshipGroups($fromdata, $siteId);
+		// Starting with the element itself.
+		$targets = [$elementId];
 
-		return $fromdata['results'];
+		// Any variants within the element, as the variant and element share the
+		// same editor pages (and can be referenced individually)
+		$targets = array_merge($targets, $this->getVariantIdsByProducts($targets));
+
+		// Find all elements that have any of these elements as targets.
+		$relationships = $this->getRelationships($targets, $siteId, true);
+
+		// Incoming connections may be coming from elements such as matrix
+		// blocks. Before retrieving proper elements and generating the map,
+		// their appropriate owner elements should be found.
+		$relationships = $this->getUsableRelationElements($relationships, $siteId);
+
+		// Retrieve the underlying elements from the relationships.
+		return $this->getElementMapData($relationships, $siteId);
 	}
 
 	/**
 	 * Retrieves a list of elements that the given element references.
-	 * @param int $elementid The ID of the element to retrieve map
+	 * @param int $elementId The ID of the element to retrieve map
 	 * information about.
 	 * @param int $siteId The ID of the site context that information should
 	 * be gathered within.
 	 */
 	public function getOutgoingElements($elementId, int $siteId)
 	{
-		if (!$elementId) {
+		if (!$elementId) { // No element, no related elements.
 			return null;
 		}
 
-		// Find outgoing relationships from this element. This includes not only direct references, but any
-		// child elements like matrix blocks must have their own external references included, this means we can
-		// check things like Entry -contains-> Matrix Block -contains-> Asset.
-		$variants = $this->getVariantsByProduct($elementId); // Element -> variants.
-		$blocks = $this->getMatrixBlocksByOwner($elementId); // Element -> matrix blocks.
-		$todata = $this->getRelationshipGroups(array_merge([$elementId], $blocks, $variants), $siteId, false); // Element + other nested content -> relationships
+		// Assemble a set of elements that should be used as the sources.
 
-		// Convert the retrieved element ids into data we use to display the map.
-		$this->processRelationshipGroups($todata, $siteId);
+		// Starting with the element itself.
+		$sources = [$elementId];
 
-		return $todata['results'];
+		// Any variants within the element, as the variant and element share the
+		// same editor pages.
+		$sources = array_merge($sources, $this->getVariantIdsByProducts($sources));
+
+		// Any matrix blocks, because they contain fields that may reference
+		// other elements
+		$sources = array_merge($sources, $this->getMatrixBlockIdsByOwners($sources));
+
+		// Any super table blocks, for the same reason as matrix blocks, and
+		// because they may themselves be contained within the matrix blocks.
+		$sources = array_merge($sources, $this->getSuperTableBlockIdsByOwners($sources));
+
+		// Any matrix blocks, again, in the case of any matrix blocks being
+		// contained within the super table blocks. This is thankfully as
+		// far as the recursion can go.
+		$sources = array_merge($sources, $this->getMatrixBlockIdsByOwners($sources));
+
+		// Find all elements that have any of these elements as sources.
+		$relationships = $this->getRelationships($sources, $siteId, false);
+
+		// Outgoing connections may be going to elements such as variants.
+		// Before retrieving proper elements and generating the map, their
+		// appropriate owner elements should be found.
+		$relationships = $this->getUsableRelationElements($relationships, $siteId);
+
+		// Retrieve the underlying elements from the relationships.
+		return $this->getElementMapData($relationships, $siteId);
 	}
 
-	private function getVariantsByProduct($element) {
-		if (!class_exists(VariantQuery::class)) { // If commerce is not installed, don't worry about variants.
+	/**
+	 * Attempts to retrieve variants for this element (used as a product), or
+	 * returns nothing if the element is not a product or if Craft Commerce is
+	 * not installed.
+	 * @param $elementIds The element(s) to retrieve variants of.
+	 * @return array An array of element IDs.
+	 */
+	private function getVariantIdsByProducts($elementIds)
+	{
+		// Make sure commerce is installed for this.
+		if (!Craft::$app->getPlugins()->getPlugin('commerce')) {
 			return [];
 		}
+
 		$conditions = [
-			'and',
-			['productId' => $element],
+			'productId' => $elementIds,
 		];
 		return (new Query())
 			->select('id')
@@ -122,91 +181,15 @@ class Renderer extends Component
 	}
 
 	/**
-	 * @param array $elementids The array of elements to get relationships for.
-	 * @param int $siteid The site ID that relationships should exist within.
-	 * @param bool $getsources Set to true when the elementids are for target elements, and the sources are being
-	 * searched for, or false when the elementids are for source elements, and the targets are being looked for.
+	 * Retrieves matrix blocks that are owned by the provided elements.
+	 * @param $elementId The element(s) to retrieve blocks for.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
 	 */
-	private function getRelationshipGroups(array $elementids, int $siteid, bool $getsources)
+	private function getMatrixBlockIdsByOwners($elementIds)
 	{
-		if ($getsources) {
-			$fromcol = 'targetId';
-			$tocol = 'sourceId';
-		} else {
-			$fromcol = 'sourceId';
-			$tocol = 'targetId';
-		}
-
-		// Get a list of elements where the given element IDs are part of the relationship,
-		// either target or source, defined by `getsources`.
 		$conditions = [
-			'and',
-			[
-				'in',
-				$fromcol,
-				$elementids
-			],
-			[
-				'or',
-				['sourceSiteId' => null],
-				['sourceSiteId' => $siteid],
-			],
-		];
-
-		$results = (new Query())
-			->select('[[r.' . $tocol . ']] AS id, [[e.type]] AS type')
-			->from('{{%relations}} r')
-			->leftJoin('{{%elements}} e', '[[r.' . $tocol . ']] = [[e.id]]')
-			->where($conditions)
-			->all();
-
-		// Create element type groups in order to further process the element list.
-		$elements = [
-			'craft\elements\MatrixBlock' => [],
-			'craft\elements\Entry' => [],
-			'craft\elements\GlobalSet' => [],
-			'craft\elements\Category' => [],
-			'craft\elements\Tag' => [],
-			'craft\elements\Asset' => [],
-			'craft\elements\User' => [],
-			'craft\commerce\elements\Product' => [],
-			'craft\commerce\elements\Variant' => [],
-			'Other' => [],
-			'results' => [],
-		];
-		$this->integrateGroupData($elements, $results);
-		return $elements;
-	}
-
-	/**
-	 * Processes input elements from a database query, and sorts them by type into an appropriate
-	 * container for further processing.
-	 * @param array &$groups A reference to the groups container to store elements within.
-	 * @param array &elements The list of elements to store within the container.
-	 */
-	private function integrateGroupData(array &$groups, array &$elements) {
-		foreach ($elements as $element) {
-			if (isset($groups[$element['type']])) { // We know the type of element this is, store it.
-				$groups[$element['type']][] = $element['id'];
-			} else { // Some kind of element not handled by the map, store it in `Other`.
-				$groups['Other'][] = $element['id'];
-			}
-		}
-	}
-
-	/**
-	 * Retrieves a list of matrix block IDs based on the given owner ids.
-	 * @param owners The owner ID to retrieve matrix blocks for.
-	 */
-	private function getMatrixBlocksByOwner($owner) {
-		$conditions = [
-			'and',
-			['ownerId' => $owner],
-			[
-				'or',
-				['ownerSiteId' => null],
-				['ownerSiteId' => Craft::$app->getSites()->currentSite->id],
-			]
+			'ownerId' => $elementIds,
 		];
 		return (new Query())
 			->select('id')
@@ -216,12 +199,209 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Finds owner elements of matrix group items, and returns those elements.
-	 * @param group A list of matrix block ids to find owners for.
+	 * Retrieves super table blocks that are owned by the provided elements.
+	 * @param $elementId The element(s) to retrieve blocks for.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
 	 */
-	private function processMatrixGroup($group) {
+	private function getSuperTableBlockIdsByOwners($elementIds)
+	{
+		// Make sure super table is installed.
+		if (!Craft::$app->getPlugins()->getPlugin('super-table')) {
+			return [];
+		}
+
 		$conditions = [
-			'mb.id' => $group,
+			'ownerId' => $elementIds,
+		];
+		return (new Query())
+			->select('id')
+			->from('{{%supertableblocks}}')
+			->where($conditions)
+			->column();
+	}
+
+	/**
+	 * Retrieves elements that are either the source or target of relationships
+	 * with the provided elements.
+	 * @param array $elementIds The array of elements to get relationships for.
+	 * @param int $siteId The site ID that relationships should exist within.
+	 * @param bool $getSources Set to true when the elementIds are for target
+	 * elements, and the sources are being searched for, or false when the
+	 * elementIds are for source elements, and the targets are being looked for.
+	 * @return array An array of arrays, the outer array being keyed by element
+	 * type, and the inner arrays containing element IDs.
+	 */
+	private function getRelationships(array $elementIds, int $siteId, bool $getSources)
+	{
+		if ($getSources) {
+			$fromcol = 'targetId';
+			$tocol = 'sourceId';
+		} else {
+			$fromcol = 'sourceId';
+			$tocol = 'targetId';
+		}
+
+		// Get a list of elements where the given element IDs are part of the relationship,
+		// either target or source, defined by `getSources`.
+		$conditions = [
+			'and',
+			[
+				'in',
+				$fromcol,
+				$elementIds
+			],
+			[
+				'or',
+				['sourceSiteId' => null],
+				['sourceSiteId' => $siteId],
+			],
+		];
+
+		$results = (new Query())
+			->select('[[e.id]] AS id, [[e.type]] AS type')
+			->from('{{%relations}} r')
+			->leftJoin('{{%elements}} e', '[[r.' . $tocol . ']] = [[e.id]]')
+			->where($conditions)
+			->all();
+
+		$results = $this->groupByType($results);
+
+		return $results;
+	}
+
+	/**
+	 * Finds elements within the relation set such as matrix blocks that should
+	 * instead reference their owning elements.
+	 * @param array $elements The elements to find usable elements for.
+	 * @param int $siteId The site that the elements should be within.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
+	 */
+	private function getUsableRelationElements(array $elements, int $siteId)
+	{
+		$results = [];
+
+		// This will iterate over available elements, bundled by type,
+		// processing whole groups by type, either adding them to the result
+		// set if they can be used outright, or retrieving a related element
+		// to use to show the relationship instead.
+		while (count($elements)) {
+			// Retrieve the next element type.
+			reset($elements);
+			$type = key($elements);
+
+			// Determine if that element type should be processed or if it
+			// should simply be added to the result set.
+			switch ($type) {
+				/* Just in case individual variant mapping turns out to be a bad idea.
+				// Variants should instead map to their products, as those are
+				// the elements through which they may be edited.
+				case 'craft\\commerce\\elements\\Variant':
+					$items = $this->getProductsForVariants($elements[$type]);
+					unset($elements[$type]);
+					$items = $this->groupByType($items);
+					$elements = $this->mergeGroups($elements, $items);
+					break;
+				*/
+				// Matrix blocks should find their owners, and then those may
+				// be reprocessed to determine if they are usable.
+				case 'craft\\elements\\MatrixBlock':
+					$items = $this->getOwnersForMatrixBlocks($elements[$type]);
+					unset($elements[$type]);
+					$items = $this->groupByType($items);
+					$elements = $this->mergeGroups($elements, $items);
+					break;
+				// Super table blocks should find their owners, and then those
+				// may be reprocessed to determine if they are usable.
+				case 'verbb\\supertable\\elements\\SuperTableBlockElement':
+					$items = $this->getOwnersForSuperTableBlocks($elements[$type]);
+					unset($elements[$type]);
+					$items = $this->groupByType($items);
+					$elements = $this->mergeGroups($elements, $items);
+					break;
+				// Anything not processed above is alright to be added to the
+				// result set and then retrieved later if it is supported.
+				default:
+					foreach ($elements[$type] as $element) {
+						$results[] = [
+							'id' => $element,
+							'type' => $type,
+						];
+					}
+					unset($elements[$type]);
+					break;
+			}
+		}
+		return $results;
+	}
+
+	/**
+	 * Sorts the elements provided into individual arrays, keyed by type.
+	 * @param array $elements Tge elements to group.
+	 * @return array An array of arrays, the outer array being keyed by element
+	 * type, and the inner arrays containing element IDs.
+	 */
+	private function groupByType(array $elements)
+	{
+		$results = [];
+		foreach ($elements as $element) {
+			if (!isset($results[$element['type']])) {
+				$results[$element['type']] = [];
+			}
+			$results[$element['type']][] = $element['id'];
+		}
+		return $results;
+	}
+
+	/**
+	 * Merges two groups in the same format as that provided by `groupByType`.
+	 * @param array $groupsA The first group to merge.
+	 * @param array $groupsB The second group to merge.
+	 * @return array The two merged groups. An array of arrays, the outer array
+	 * being keyed by element type, and the inner arrays containing element IDs.
+	 */
+	private function mergeGroups(array $groupsA, array $groupsB)
+	{
+		foreach ($groupsB as $type => $elements) {
+			if (!isset($groupsA[$type])) {
+				$groupsA[$type] = [];
+			}
+			$groupsA[$type] = array_merge($groupsA[$type], $elements);
+		}
+		return $groupsA;
+	}
+
+	/**
+	 * Retrieves product IDs/types of product elements using the variant IDs
+	 * provided.
+	 * @param $elementIds An array of IDs.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
+	 */
+	private function getProductsForVariants(array $elementIds)
+	{
+		$conditions = [
+			'productId' => $elementIds,
+		];
+		return (new Query())
+			->select('id')
+			->from('{{%commerce_variants}}')
+			->where($conditions)
+			->column();
+	}
+
+	/**
+	 * Retrieves owner IDs/types of owning elements using the matrix block IDs
+	 * provided.
+	 * @param $elementIds An array of IDs.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
+	 */
+	private function getOwnersForMatrixBlocks($elementIds)
+	{
+		$conditions = [
+			'mb.id' => $elementIds,
 		];
 		return (new Query())
 			->select('[[e.id]] AS id, [[e.type]] AS type')
@@ -232,12 +412,34 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts entries into a list of standardized result items.
-	 * @param group The IDs of the entries to convert.
+	 * Retrieves owner IDs/types of owning elements using the super table block
+	 * IDs provided.
+	 * @param $elements An array of IDs.
+	 * @return array An array of elements, with their ID `id` and element type
+	 * `type`.
 	 */
-	private function processEntryGroup($group, $siteId) {
+	private function getOwnersForSuperTableBlocks($elementIds)
+	{
+		$conditions = [
+			'stb.id' => $elementIds,
+		];
+		return (new Query())
+			->select('[[e.id]] AS id, [[e.type]] AS type')
+			->from('{{%supertableblocks}} stb')
+			->leftJoin('{{%elements}} e', '[[stb.ownerId]] = [[e.id]]')
+			->where($conditions)
+			->all();
+	}
+
+	/**
+	 * Retrieves entry elements based on a set of IDs.
+	 * @param $elementIds The IDs of the entries to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
+	 */
+	private function getEntryElements($elementIds, $siteId)
+	{
 		$criteria = new EntryQuery('craft\elements\Entry');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$criteria->siteId = $siteId;
 		$elements = $criteria->all();
 
@@ -254,12 +456,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts global set items into a list of standardized result items.
-	 * @param group The IDs of the global sets to convert.
+	 * Retrieves globalset elements based on a set of IDs.
+	 * @param $elementIds The IDs of the globalsets to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processGlobalSetGroup($group) {
+	private function getGlobalSetElements($elementIds, $siteId)
+	{
 		$criteria = new GlobalSetQuery('craft\elements\GlobalSet');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$elements = $criteria->all();
 
 		$results = [];
@@ -275,12 +479,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts categories into a list of standardized result items.
-	 * @param group The IDs of the categories to convert.
+	 * Retrieves category elements based on a set of IDs.
+	 * @param $elementIds The IDs of the categories to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processCategoryGroup($group, $siteId) {
+	private function getCategoryElements($elementIds, $siteId)
+	{
 		$criteria = new CategoryQuery('craft\elements\Category');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$criteria->siteId = $siteId;
 		$elements = $criteria->all();
 
@@ -297,12 +503,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts tags into a list of standardized result items.
-	 * @param group The IDs of the tags to convert.
+	 * Retrieves tag elements based on a set of IDs.
+	 * @param $elementIds The IDs of the tags to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processTagGroup($group) {
+	private function getTagElements($elementIds, $siteId)
+	{
 		$criteria = new TagQuery('craft\elements\Tag');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$elements = $criteria->all();
 
 		$results = [];
@@ -318,12 +526,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts assets into a list of standardized result items.
-	 * @param group The IDs of the assets to convert.
+	 * Retrieves asset elements based on a set of IDs.
+	 * @param $elementIds The IDs of the assets to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processAssetGroup($group) {
+	private function getAssetElements($elementIds, $siteId)
+	{
 		$criteria = new AssetQuery('craft\elements\Asset');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$elements = $criteria->all();
 
 		$results = [];
@@ -339,12 +549,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts users into a list of standardized result items.
-	 * @param group The IDs of the users to convert.
+	 * Retrieves user elements based on a set of IDs.
+	 * @param $elementIds The IDs of the users to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processUserGroup($group) {
+	private function getUserElements($elementIds, $siteId)
+	{
 		$criteria = new UserQuery('craft\elements\User');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$elements = $criteria->all();
 
 		$results = [];
@@ -360,15 +572,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts products into a list of standardized result items.
-	 * @param group The IDs of the products to convert.
+	 * Retrieves product elements based on a set of IDs.
+	 * @param $elementIds The IDs of the products to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processProductGroup($group, $siteId) {
-		if (!class_exists(ProductQuery::class)) { // Commerce not installed.
-			return [];
-		}
+	private function getProductElements($elementIds, $siteId)
+	{
 		$criteria = new ProductQuery('craft\commerce\elements\Product');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$criteria->siteId = $siteId;
 		$elements = $criteria->all();
 
@@ -385,15 +596,14 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Converts variants into a list of standardized result items.
-	 * @param group The IDs of the variants to convert.
+	 * Retrieves variant elements based on a set of IDs.
+	 * @param $elementIds The IDs of the variants to retrieve.
+	 * @param $siteId The ID of the site to use as the context for element data.
 	 */
-	private function processVariantGroup($group, $siteId) {
-		if (!class_exists(VariantQuery::class)) { // Commerce not installed.
-			return [];
-		}
+	private function getVariantElements($elementIds, $siteId)
+	{
 		$criteria = new VariantQuery('craft\commerce\elements\Variant');
-		$criteria->id = $group;
+		$criteria->id = $elementIds;
 		$criteria->siteId = $siteId;
 		$elements = $criteria->all();
 
@@ -410,56 +620,28 @@ class Renderer extends Component
 	}
 
 	/**
-	 * Iterates over elements within each group, converting what it can find into result sets.
-	 * @param groups A reference to the groups container that contains the processed and unprocessed elements.
-	 * @param siteId The ID of the site to retrieve elements within.
+	 * Converts a set of elements to an array of map-ready associative arrays.
+	 * @param array $elements An array of elements (with `id` and `type`) to
+	 * retrieve map information for.
+	 * @param int $siteId The ID of the site to retrieve element data within.
+	 * @return array A set of elements that can be used to display the map.
 	 */
-	private function processRelationshipGroups(&$groups, $siteId) {
-		if (count($groups['craft\elements\MatrixBlock'])) {
-			$data = $this->processMatrixGroup($groups['craft\elements\MatrixBlock']); // Process the data for this group.
-			$groups['craft\elements\MatrixBlock'] = []; // Clear the data for this group.
-			$this->integrateGroupData($groups, $data); // Re-integrate new data into the group container.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\commerce\elements\Product'])) {
-			$data = $this->processProductGroup($groups['craft\commerce\elements\Product'], $siteId); // Process the data for this group.
-			$groups['craft\commerce\elements\Product'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\commerce\elements\Variant'])) {
-			$data = $this->processVariantGroup($groups['craft\commerce\elements\Variant'], $siteId); // Process the data for this group.
-			$groups['craft\commerce\elements\Variant'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\Entry'])) {
-			$data = $this->processEntryGroup($groups['craft\elements\Entry'], $siteId); // Process the data for this group.
-			$groups['craft\elements\Entry'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\GlobalSet'])) {
-			$data = $this->processGlobalSetGroup($groups['craft\elements\GlobalSet']); // Process the data for this group.
-			$groups['craft\elements\GlobalSet'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\Category'])) {
-			$data = $this->processCategoryGroup($groups['craft\elements\Category'], $siteId); // Process the data for this group.
-			$groups['craft\elements\Category'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\Tag'])) {
-			$data = $this->processTagGroup($groups['craft\elements\Tag']); // Process the data for this group.
-			$groups['craft\elements\Tag'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\Asset'])) {
-			$data = $this->processAssetGroup($groups['craft\elements\Asset']); // Process the data for this group.
-			$groups['craft\elements\Asset'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
-		} else if (count($groups['craft\elements\User'])) {
-			$data = $this->processUserGroup($groups['craft\elements\User']); // Process the data for this group.
-			$groups['craft\elements\User'] = []; // Clear the data for this group.
-			$groups['results'] = array_merge($groups['results'], $data); // Add the results to the set.
-			$this->processRelationshipGroups($groups, $siteId); // Process more groups.
+	private function getElementMapData(array $elements, int $siteId)
+	{
+		$elements = $this->groupByType($elements);
+		$results = [];
+
+		while (count($elements)) {
+			// Retrieve the next element type.
+			reset($elements);
+			$type = key($elements);
+
+			if (isset(self::ELEMENT_TYPE_MAP[$type])) {
+				$results = array_merge($results, call_user_func([$this, self::ELEMENT_TYPE_MAP[$type]], $elements[$type], $siteId));
+			}
+
+			unset($elements[$type]);
 		}
+		return $results;
 	}
 }